requirements.txt
Flask==3.1.1
Flask_MySQLdb==2.0.0
Pillow==11.3.0
reportlab==4.4.3
